% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFOsalmonids-package.R
\docType{package}
\name{DFOsalmonids-package}
\alias{DFOsalmonids}
\alias{DFOsalmonids-package}
\title{DFOsalmonids: DFOsalmonids}
\description{
Example data containing salmonid catches for DFO stats course.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/DFOsalmonids}
}

}
\author{
\strong{Maintainer}: Colton Stephens \email{coltonstephens@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-8744-6405}{ORCID})

}
\keyword{internal}
