% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-age.R
\docType{data}
\name{scale_age}
\alias{scale_age}
\title{Scale Age Data}
\format{
\subsection{\code{scale_age}}{

A data frame with 112 rows and 3 columns:
\describe{
\item{brood_year}{Brood year}
\item{captures}{Number of samples in age group}
\item{age}{Age of fish based on scale sample}
}
}
}
\usage{
scale_age
}
\description{
Example data for counts of salmonids from different age classes based on
scale data.
}
\keyword{datasets}
