% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFOsalmonids-package.R
\docType{package}
\name{DFOsalmonids-package}
\alias{DFOsalmonids}
\alias{DFOsalmonids-package}
\title{DFOsalmonids: Example data for the DFO sample statistics course}
\description{
Example data containing theoretical catches of salmonids from different age classes as well as tagged catch/release data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/DFOsalmonids}
}

}
\author{
\strong{Maintainer}: Colton Stephens \email{coltonstephens@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-8744-6405}{ORCID})

}
\keyword{internal}
